Put current models in new Protocol3 class.
