+++ lib/timex_datalink_client/protocol_3/end.rb
@@ -1,0 +1,10 @@
+# frozen_string_literal: true
+
+require "timex_datalink_client/end"
+
+class TimexDatalinkClient
+  class Protocol3
+    class End < End
+    end
+  end
+end
