+++ lib/timex_datalink_client/protocol_3/sync.rb
@@ -1,0 +1,10 @@
+# frozen_string_literal: true
+
+require "timex_datalink_client/sync"
+
+class TimexDatalinkClient
+  class Protocol3
+    class Sync < Sync
+    end
+  end
+end
