+++ spec/lib/protocol_3/timex_datalink_client/eeprom/list_spec.rb
@@ -2,7 +2,7 @@
 
 require "spec_helper"
 
-describe TimexDatalinkClient::Eeprom::List do
+describe TimexDatalinkClient::Protocol3::Eeprom::List do
   let(:list_entry) { "Muffler Bearings" }
   let(:priority) { 0 }
