+++ spec/lib/protocol_3/timex_datalink_client/notebook_adapter_spec.rb
