+++ spec/lib/protocol_3/timex_datalink_client/sound_options_spec.rb
@@ -2,7 +2,7 @@
 
 require "spec_helper"
 
-describe TimexDatalinkClient::SoundOptions do
+describe TimexDatalinkClient::Protocol3::SoundOptions do
   let(:hourly_chime) { false }
   let(:button_beep) { false }
